refine the scripts directory structure
