+++ scripts/include/debug/debug.h
@@ -1,0 +1,1 @@
+void check_struct_offset(void);
