+++ scripts/include/lib/lib.h
@@ -1,0 +1,2 @@
+#define offsetof(type, member) (size_t)&(((type*)0)->member)
+
