+++ scripts/parse_boot_img/main.c
@@ -1,10 +1,3 @@
-/*************************************************************************
-    > File Name: scripts/parse_boot_img/main.c
-    > Author:		
-    > Mail:		
-    > Created Time: 2021年05月22日 星期六 11时13分38秒
- ************************************************************************/
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <getopt.h>
@@ -16,69 +9,21 @@
 #include <sys/mman.h>
 #include <unistd.h>
 #include <stdbool.h>
+#include <libgen.h>
+#include "type.h"
+#include "debug.h"
 
-enum file_attr {
-	ATTR_READ_ONLY = 0x01,
-	ATTR_HIDDEN = 0x02,
-	ATTR_SYSTEM = 0x04,
-	ATTR_VOLUME_ID = 0x08,
-	ATTR_DIRECTORY = 0x10,
-	ATTR_ARCHIVE = 0x20,
-	ATTR_LONG_NAME = ATTR_READ_ONLY | ATTR_HIDDEN | \
-			ATTR_SYSTEM | ATTR_VOLUME_ID,
-};
-
-#pragma pack(1)
-struct root_dir_region {
-	char dir_name[11];
-	unsigned char dir_attr;
-	unsigned char dir_ntres;
-	unsigned char dir_crt_time_tenth;
-	unsigned short dir_crt_time;
-	unsigned short dir_crt_date;
-	unsigned short dir_lst_acc_date;
-	unsigned short dir_fst_clus_hi; // High word of this entry’s first cluster number (always 0 for a FAT12 or FAT16 volume).
-	unsigned short dir_wrt_time;
-	unsigned short dir_wrt_date;
-	unsigned short dir_fst_clus_lo; // Low word of this entry’s first cluster number.
-	unsigned int dir_file_size;
-};
-
-struct file_allocation_table {
-	unsigned short jmp;
-	char hi_start;
-	char oem_name[8];
-	unsigned short byte_per_sec;
-	unsigned char sec_per_clus;
-	unsigned short rsvd_sec_cnt;
-	unsigned char num_fats;
-	unsigned short root_ent_cnt;
-	unsigned short tot_sec16;
-	unsigned char media_type;
-	unsigned short fat_sz16;
-	unsigned short sec_per_trk;
-	unsigned short num_heads;
-	unsigned int hidd_sec;
-	unsigned int tot_sec32;
-	unsigned char drv_num;
-	unsigned char reserved1;
-	unsigned char boot_sig;
-	unsigned int vol_id;
-	char vol_lab[11];
-	char file_sys_type[8];
-};
-#pragma pack()
-
-#define SIZE_PER_SEC 512
-#define BYTES_PER_LINE 16
-
+/*TODO:
+ * 1. determine the -f is fat12/fat16/fat32
+ * 2. re-work the struct of file_allocation_table to support fat32
+ */
 static char *progname;
-static unsigned long file_size = 1 * SIZE_PER_SEC;
-static struct file_allocation_table fat;
+unsigned long file_size = 1 * SIZE_PER_SEC;
+struct file_allocation_table fat;
 /*  the middle two bytes are the fat.jmp code start address */
-static unsigned short *text_start;
-static char file_name[256];
-static bool get_file = false;
+unsigned short *text_start;
+char file_name[256];
+bool get_file = false;
 
 static struct option long_options[] = {
 	{"show hex dump",        no_argument,       0, 's'},
@@ -86,15 +31,13 @@ static struct option long_options[] = {
 	{"get/show boot sector", required_argument, 0, 'p'},
 };
 
-#define offsetof(type, member) (size_t)&(((type*)0)->member)
-
 static void usage(void)
 {
 	printf("Usage: %s <-f file_name> [-sp]\n", progname);
 	exit(-1);
 }
 
-static int to_ascii(char *str, int num)
+static int to_ascii(const char *str, int num)
 {
 	int i, ofset = 0;
 	char buf[num + 3];
@@ -159,7 +102,7 @@ static int show_hex_dump(char *arg)
 {
 	int line = file_size / BYTES_PER_LINE;
 	int i, offset = 0, *start, fd, ret;
-	char *tmp_char;
+	const char *tmp_char;
 	unsigned short *tmp_data;
 
 	start = attach_file(arg, &fd, offset);
@@ -168,7 +111,7 @@ static int show_hex_dump(char *arg)
 	}
 	for (int i = 0; i < line; i ++) {
 		tmp_data = (unsigned short *)((char  *)start + offset);
-		tmp_char = (char *)tmp_data;
+		tmp_char = (const char *)tmp_data;
 		printf("0x%.8x   %.4x %.4x %.4x %.4x %.4x %.4x %.4x %.4x  ",
 		offset, *tmp_data, *(tmp_data+1), *(tmp_data+2), *(tmp_data+3),
 		*(tmp_data+4), *(tmp_data+5), *(tmp_data+6), *(tmp_data+7));
@@ -215,10 +158,10 @@ static void show_boot_struct()
 {
 	/* the jmp used the first byte, skip the fist byte. The start address is the second byte for low byte,
 	* and the third byte for higher byte */
-	unsigned short low, hi;
+	unsigned long low, hi;
 	hi = (fat.hi_start & 0xff) << 8;
 	low = (fat.jmp & (~0xff)) >> 8;
-	text_start = (unsigned short *)(hi | low);
+	text_start = (short unsigned int *)(hi | low);
 	fprintf(stdout, "BootSector:\n"
 		"jmp              %p\n"
 		"BS_OEMName:      %s\n"
@@ -247,52 +190,30 @@ static void show_boot_struct()
 		fat.vol_id, fat.vol_lab, fat.file_sys_type);
 
 }
-
-void check_struct_offset(void)
+#if 0
+static int get_file_type()
 {
 
-	fprintf(stdout, "Check struct offset:\n"
-		"BS_OEMName:      %ld\n"
-		"BPB_BytesPerSec: %ld\n"
-		"BPB_SecPerClus:  %ld\n"
-		"BPB_RsvdSecCnt:  %ld\n"
-		"BPB_NumFATs:     %ld\n"
-		"BPB_RootEntCnt:  %ld\n"
-		"BPB_TotSec16:    %ld\n"
-		"BPB_Media:       %ld\n"
-		"BPB_FATSz16:     %ld\n"
-		"BPB_SecPerTrk:   %ld\n"
-		"BPB_NumHeads:    %ld\n"
-		"BPB_HiddSec:     %ld\n"
-		"BPB_TotSec32:    %ld\n"
-		"BS_DrvNum:       %ld\n"
-		"BS_Reserved1:    %ld\n"
-		"BS_BootSig:      %ld\n"
-		"BS_VolID:        %ld\n"
-		"BS_VolLab:       %ld\n"
-		"BS_FileSysType:  %ld\n",
-		offsetof(struct file_allocation_table, oem_name), offsetof(struct file_allocation_table, byte_per_sec),
-		offsetof(struct file_allocation_table, sec_per_clus), offsetof(struct file_allocation_table, rsvd_sec_cnt),
-		offsetof(struct file_allocation_table, num_fats), offsetof(struct file_allocation_table, root_ent_cnt),
-		offsetof(struct file_allocation_table, tot_sec16), offsetof(struct file_allocation_table, media_type),
-		offsetof(struct file_allocation_table, fat_sz16), offsetof(struct file_allocation_table, sec_per_trk),
-		offsetof(struct file_allocation_table, num_heads), offsetof(struct file_allocation_table, hidd_sec),
-		offsetof(struct file_allocation_table, tot_sec32), offsetof(struct file_allocation_table, drv_num),
-		offsetof(struct file_allocation_table, reserved1), offsetof(struct file_allocation_table, boot_sig),
-		offsetof(struct file_allocation_table, vol_id), offsetof(struct file_allocation_table, vol_lab),
-		offsetof(struct file_allocation_table, file_sys_type));
+	int fat_sz, tot_sec, data_clus_cnt;
+	if (fat.fat_sz16 != 0) {
+		fat_sz = fat.fat_sz16;
+	} else {
+		fat_sz = fat.fat_sz32;
+	}
 
 }
+#endif
 
 int main (int argc, char *argv[])
 {
+	char c;
+	int opt_index = 0, ret = 0;
+	const char *optstring = "sp:f:";
+	progname = basename(argv[0]);
 	if (argc < 2) {
 		usage();
 	}
 
-	char c;
-	int opt_index = 0, ret = 0;
-	const char *optstring = "sp:f:";
 	while ((c = getopt_long(argc, argv, optstring, long_options,
 		&opt_index)) != -1) {
 		switch(c) {
